Shuffle around some tests
