+++ openff/models/models.py
@@ -1,7 +1,7 @@
 from openff.units import Quantity
 
 from openff.models._pydantic import BaseModel
-from openff.models.types import custom_quantity_encoder
+from openff.models.types.serialization import custom_quantity_encoder
 
 
 class DefaultModel(BaseModel):
